requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-dotenv>=1.0.1
requests>=2.31.0
openai>=1.12.0
python-multipart>=0.0.9
pydantic>=2.7.2,<3.0.0
sseclient-py>=1.8.0
fastapi-mcp>=0.1.0
boto3>=1.34.0
jinja2>=3.1.0
slack-bolt>=1.18.0 